% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{flatten_named_matrix}
\alias{flatten_named_matrix}
\title{Flatten a Named Matrix to a Named Vector}
\usage{
flatten_named_matrix(x, sep = "_")
}
\arguments{
\item{x}{A matrix with non-empty row and column names. The input should be a valid matrix.}

\item{sep}{A character string used as the delimiter between row and column names in the constructed names.
Defaults to \code{"_"}}
}
\value{
A vector representing the matrix elements with names formatted as "\if{html}{\out{<row>}}\if{html}{\out{<sep>}}\if{html}{\out{<column>}}".
}
\description{
Converts a matrix with row and column names into a named vector. The names in the resulting vector are
constructed by concatenating the corresponding row and column names using a specified separator.
}
\details{
The conversion preserves the column-major order inherent in R matrices. For example, if the
matrix is constructed with \code{byrow = FALSE}, the resulting vector will list all elements of the first
column (with names based on the corresponding rows and that column name), followed by the second column, etc.
}
\examples{
mx <- matrix(
  c(0, 5, 3, 1), ncol = 2L,
  dimnames = list(c("present", "absent"), c("colorectal", "melanoma"))
)
flatten_named_matrix(mx)
}
