% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyenrich.R
\name{list_to_contingency_tables}
\alias{list_to_contingency_tables}
\title{Convert a Named List of Vectors into Contingency Tables}
\usage{
list_to_contingency_tables(input_list)
}
\arguments{
\item{input_list}{A named list where each element is a character or factor vector.}
}
\value{
A named list of contingency tables (one per unique element).
Each contingency table shows counts of presence and absence of the element across groups.
Each table is a numeric matrix
}
\description{
This function takes a named list of character or factor vectors
and produces a list of contingency tables. Each table tests for
the presence or absence of a specific unique element across the groups.
}
\examples{
input <- list(
  colorectal = c("APC", "APC", "APC", "TP53", "APC"),
  melanoma = c("BRAF", "BRAF", "BRAF", "TP53"),
  breast = c("BRAF", "BRAF", "BRAF", "TP53")
)
list_to_contingency_tables(input)

}
