% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{plot_rainforest}
\alias{plot_rainforest}
\title{Create rainforest plot}
\usage{
plot_rainforest(data, max_odds = 5)
}
\arguments{
\item{data}{a data.frame created by \code{\link[=contingency_tables_to_fisher]{contingency_tables_to_fisher()}}. Must include the columns "element", "p.value", "odds_ratio", "conf_level", "conf.int.lower", "conf.int.upper", "fdr".}

\item{max_odds}{the x axis upper limit. Any odds ratios > max_odds will be depicted as an open square off the chart. Any confidence interval > max_odds will have an arrow pointing off-chart}
}
\value{
ggplot
}
\description{
Create a rainforest plot showing odds ratios + confidence intervals sorted based on significance
}
\examples{
#Step 1: Create a list of two or more vectors.
input <- list(
  colorectal = c("APC", "APC", "APC", "TP53", "APC"),
  melanoma = c("BRAF", "BRAF", "BRAF", "BRAF", "BRAF", "TP53", "APC")
)

# Step 2: Create contingency tables
contingency_tables <- list_to_contingency_tables(input)

# Step 3: Compute fisher p values & odds ratios
comparison <- contingency_tables_to_fisher(contingency_tables)

# Step 5: Visualise Results
plot_rainforest(comparison)
}
